+++ Services/Interfaces/ICustomerInformationService.cs
@@ -6,7 +6,7 @@ namespace ChallengeAtmApi.Services.Interfaces
     {
         Task<CustomerInformation> GetCustomerByCardNumber(int cardNumber);
         Task<CustomerInformation> GetCustomerById(Guid id);
-        Task<CustomerInformation> AddAmmountToBalance(Guid id, float amount);
+        Task<CustomerInformation> AddAmmountToBalance(Guid id, double amount);
 
     }
 }
