+++ Services/Interfaces/ITransactionHistoryService.cs
@@ -8,6 +8,6 @@ namespace ChallengeAtmApi.Services.Interfaces
         //Task<TransactionHistory> DoTransactionWithdrawAsync(Guid customerId, float ammount);
         Task<TransactionCheckDto> DoTransactionCheckAsync(int cardNumber);
         //Task<TransactionHistory> DoTransactionTransferAsync(Guid customerId, float ammout, Guid destinationCustomerId);
-        Task<TransactionHistory> DoTransactionDepositAsync(Guid customerId, float ammount);
+        Task<TransactionHistory> DoTransactionDepositAsync(Guid customerId, double ammount);
     }
 }
